AuthData UserData перемещены в type
